Also relocate getRelayEventKey
